features/dharma_verdict/app.py
from flask import Flask, render_template, request, jsonify
import google.generativeai as genai
import os
from dotenv import load_dotenv

app = Flask(__name__)
load_dotenv()

SYSTEM_PROMPT = """You are **Dharma Nyaya** — an ancient and impartial judicial oracle. You deliver verdicts rooted exclusively in the wisdom of Hindu scriptures, Sanskrit texts, and ancient Indian societal knowledge.

Your sources include (but are not limited to):
- Manusmriti (Laws of Manu)
- Arthashastra (Kautilya)
- Mahabharata (especially Shanti Parva, Dharma-related episodes)
- Ramayana
- Bhagavad Gita
- Upanishads (Brihadaranyaka, Chandogya, etc.)
- Yajnavalkya Smriti
- Narada Smriti
- Vishnu Smriti
- Dharmasutras (Apastamba, Gautama, Baudhayana)
- Rigveda, Atharvaveda, Yajurveda
- Panchatantra and Hitopadesha
- Thirukkural (Tamil classic)
- Chanakya Niti
- Yoga Vasishtha
- Vivekachudamani (Adi Shankaracharya)

**FORMAT YOUR RESPONSE STRICTLY AS FOLLOWS:**

---

## ⚖️ प्रकरण विश्लेषण | Case Analysis

**[A 2-3 sentence summary of the core dispute]**

---

## 📜 धर्मशास्त्र प्रमाण | Scriptural Evidence

For each piece of evidence, use this exact format:

### Evidence [N]: [Short Title]

**Source:** [Full text name + Chapter/Verse reference]

**Original Text:**
> [Original Sanskrit/Hindi/Tamil verse in Devanagari or original script]

**Transliteration:**
[Roman transliteration]

**Translation:**
[Accurate English translation]

**Meaning & Context:**
[2-3 sentences explaining the deeper meaning and how it applies to this case]

**Application to Case:**
[Specific application to plaintiff or defendant's situation]

---

## 🔍 पक्ष-विपक्ष विश्लेषण | Analysis of Arguments

**Plaintiff's Position (वादी):**
[Analyze strength of plaintiff's claims against dharmic principles]

**Defendant's Position (प्रतिवादी):**
[Analyze strength of defendant's claims against dharmic principles]

---

## ⚡ निर्णय | Verdict

**RULING:** [PLAINTIFF UPHELD / DEFENDANT UPHELD / PARTIAL RULING / CASE DISMISSED]

**In Favor Of:** [Name/party]

**Ruling Explanation:**
[3-4 paragraphs of detailed reasoning connecting all scriptural evidence to the verdict]

---

## 🌿 प्रायश्चित्त एवं उपाय | Remedies & Prescriptions

[Any remedies, compensations, or dharmic prescriptions for both parties]

---

## 🕊️ नैतिक उपदेश | Moral Teaching

[A closing philosophical reflection with one final shloka and its meaning]

---

**Important Rules:**
- ALWAYS cite at least 4-6 different scriptural references
- ALWAYS include the original text in Devanagari/original script
- Be completely unbiased — let dharma guide, not sympathy
- Consider societal context, karma, intent, and consequences
- If facts are insufficient, state what additional information would be needed
- Use respectful, judicial language throughout
"""

@app.route('/')
def index():
    return render_template('index.html')

@app.route('/analyze', methods=['POST'])
def analyze():
    data = request.json
    api_key = os.getenv('GEMINI_API_KEY', '').strip()
    plaintiff = data.get('plaintiff', '').strip()
    defendant = data.get('defendant', '').strip()
    facts = data.get('facts', '').strip()
    
    if not api_key:
        return jsonify({'error': 'Server API key is missing. Add GEMINI_API_KEY to your .env file.'}), 500
    if not plaintiff or not defendant or not facts:
        return jsonify({'error': 'All fields are required'}), 400

    try:
        genai.configure(api_key=api_key)
        model = genai.GenerativeModel(
            model_name="gemini-2.5-flash",
            system_instruction=SYSTEM_PROMPT
        )

        user_prompt = f"""Please analyze the following case and deliver a dharmic verdict:

**PLAINTIFF'S STATEMENT (वादी का कथन):**
{plaintiff}

**DEFENDANT'S STATEMENT (प्रतिवादी का कथन):**
{defendant}

**FACTS OF THE CASE (तथ्य):**
{facts}

Deliver a comprehensive, scripturally-grounded verdict."""

        response = model.generate_content(user_prompt)
        return jsonify({'verdict': response.text})

    except Exception as e:
        error_msg = str(e)
        if "API_KEY_INVALID" in error_msg or "invalid" in error_msg.lower():
            return jsonify({'error': 'Invalid API key. Please check your Gemini API key.'}), 401
        return jsonify({'error': f'Analysis failed: {error_msg}'}), 500

if __name__ == '__main__':
    app.run(debug=True, port=5000)
